Revise TM #2 up to proper_prefix
At this point, we have to revisit the potential confusion from the
previous try. The main changes are the removal of the representation of
silent divergence in trace prefixes, which makes their unobservability
via prefixes direct; and the change from finite and infinite traces to
finite and infinite executions, still preserving the names to make
proper changes apparent at this stage.