Try to adjust the trace model to tstuck
Up to proper prefixes, everything seems to work well with some tweaks
that come back to bite us later. In particular, embedding a finite
prefix into a trace and friends are not as simple any more: before, the
single termination event was used, whereas now there one of the two
possible termination events must be ignored in the embedding, which
in turn falsifies some of the lemmas that rely on this encoding.