Unified stop parameterized by cause (I)
For a more homogeneous treatment, we return to the original
representation that uses a single constructor for finite traces, and
enrich it with an inductive datatype that represents termination
information, e.g., normal termination, execution error, etc.

As before, we can easily move back and forth between representations
until we need to reason about embeddings and proper prefixes, where
similar issues arise. These are what we need to address next.