"Complete" adjustment of trace model
The end state hack reoccurs once, as in the previous version. Instead of
leaving the two problematic results active, we abort them and see what
breaks in the criteria, and whether the commented results are really the
ones we need, or if, better yet, we have enough information in the
proofs to work around them.