Begin to revamp trace embeddings
Previously, finite trace embeddings took a list of events and injected
it into a stopped trace. As there was exactly one possible way to end a
finite trace, this worked out just fine. However, as we add multiple
trace terminators, this no longer holds and we need to adjust our
definitions and the proofs that depend on them to account for this
change. Typically, this involves parameterizing on the end event of a
trace; sometimes, we can pick an end event irrespective of the situation
(we should try to make these cases cleaner later on).

Co-authored-by: Jérémy Thibault <jeremy.thibault@ens-rennes.fr>