Start a new folder with new architecture that should be more modular
It uses Coq modules. The main goal is to allow one to write different instances of compiler between languages that have different trace model themselves. Hence, languages are parametrized by their traces, and compilation chains by traces and the two languages.

I am not sure this is the most efficient way of modularizing, but that's a start. Feel free to move parts around and change things. Also, if there is a better way instead of using modules (they are quite awkward to use when there is many level of dependencies), we could change that.