Reorganize "target++" language core
In this particular setting of two essentially identical languages, in
which the target has the ability to generate additional events, it will
make things easier to organize both languages as specializations of a
shared core.

In addition to this, function calls will be made part of core
expressions and paired with some additional restrictions expressed
externally -- some of which were already envisioned in any case. This
flat vision further simplifies recursive definitions, which otherwise
can become technically more cumbersome, as the layers need to be
combined and substructural properties bookkept.