Begin to define source and target language models
The target is the full core enveloped in the constructs of the language
model. The source is a refinement of the core, to which we add the
restriction that no Out commands (and therefore events) are allowed,
thus making these target-only.

The interesting remaining bit is the encoding of this property on
function maps and its effect on their definition: the usual fare.