Structural properties and program specification
Most of the properties of interest are standard inductive properties
on the structure on expressions which prevent certain nodes from
appearing in them.

In linking, some of these properties need to be composed, and notions of
interface, exported functions, etc., make their appearance. One should
not expect, however, that all program-context pairs be linkable.