Sets as interfaces, closedness, refined linking
To represent program "interfaces" as sets of functions used by the main
expression (noting that at the moment functions cannot call other
functions), we instantiate FSets in the same way as FMaps, sharing
constraints for interoperability. (It would be possible to use the newer
MSets, but not so for maps.)

Based on this we determine that a full program is closed iff the set of
functions used in the main expression is exactly the set of functions
defined, jointly, by the partial program and the context. Of course,
this approximation can be refined in many manners.

Partial program-context linking is now defined by cases, given that the
operation must succeed unconditionally and we do not wish to further
complicate the representation of both programs and contexts. There is
consequently a bad case where the linking operation produces garbage,
which should be ruled out by adequate assumptions whenever relevant.