Harmonize trace model
To keep the traces and evaluation closer to one another, model traces
as list of events, and define those separately -- together with a few
simple properties, which are now qualified by traced produced by the
semantics.

In addition, while rewriting the relation between traces for the new
model, fix a typo in it that was yet undiscovered by the partial state
of the proofs.