Restore use of trace relation
Using a regular list of events as a trace makes it easier to connect
the results of evaluation at the various level in the proofs, but
obscures the structure of the traces which was previously directly
inferred from the type of traces.

Consequently, now we reason specifically about traces generated by the
semantics, appropriately characterized, and restore the trace relation
between traces and "clean" traces, used by the main result.

This is the last interesting bit in the formalization, not considering
some technical lemmas about the well-formedness of various
transformations, whose current use is not controlled due to the quirks
of the language model.

The time is right to explore variations of the scheme as it stands.