Add robust example to build
This series of commits adapts the robust compilation example, where the
target can generate more events than the source, to the purely vanilla
setting of the language and compiler chain models in the other examples,
and brings it a little closer to the paper proof.

The main trick is to unpack the core language into two separate
languages, source and target, making their differences explicit in the
definitions of the data types themselves, with the disadvantage of
repetition. What we gain is the ability to express the compilation chain
cleanly in the framework while retaining the simplicity of the proofs.

There is exactly one place in the proof where it is almost unavoidable
to appeal to an external notion of well-formedness which cannot be
easily packed into the existing definitions -- or these unpacked to
include it. It is not pretty, in any case, though it is also not very
intrusive.

In addition, there is an unproved helper on a property of standard weak
maps that merely requires a bit of digging in the standard library.

As can be seen, the proofs are repetitive but simple and can be
expressed much more compactly with a bit of Ltac.